requirements.txt
astroid==2.4.2
autopep8==1.5.4
cffi==1.14.3
click==8.0.1
cryptography==3.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.9
pep517==0.10.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dotenv==0.18.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
